+++ lib/frontend/project-sidebar-internal/src/Header/SidebarHeader.test.ts
@@ -24,14 +24,13 @@ import { describe, it, expect, vi } from "vitest";
 import { shallowMount } from "@vue/test-utils";
 import SidebarHeader from "./SidebarHeader.vue";
 import { example_config } from "../project-sidebar-example-config";
-import { ref } from "vue";
 import * as strict_inject from "@tuleap/vue-strict-inject";
 
 vi.mock("@tuleap/vue-strict-inject");
 
 describe("SidebarHeader", () => {
     it("displays the sidebar header", () => {
-        vi.spyOn(strict_inject, "strictInject").mockReturnValue(ref(example_config));
+        vi.spyOn(strict_inject, "strictInject").mockReturnValue(example_config);
         const wrapper = shallowMount(SidebarHeader, {
             propsData: {
                 is_sidebar_collapsed: false,
@@ -44,7 +43,7 @@ describe("SidebarHeader", () => {
     it("does not display the administration link when the user is not a project administrator", () => {
         const config = example_config;
         config.user.is_project_administrator = false;
-        vi.spyOn(strict_inject, "strictInject").mockReturnValue(ref(config));
+        vi.spyOn(strict_inject, "strictInject").mockReturnValue(config);
         const wrapper = shallowMount(SidebarHeader, {
             propsData: {
                 is_sidebar_collapsed: false,